REFA-0001 - refactor packages for retrofit providers and preferences providers
