+++ core/src/main/java/ru/easycode/words504/core/data/MakeService.kt
@@ -1,6 +1,6 @@
 package ru.easycode.words504.core.data
 
-import ru.easycode.words504.data.ProvideRetrofitBuilder
+import ru.easycode.words504.data.providers.retrofit.ProvideRetrofitBuilder
 
 interface MakeService {
