+++ core/src/main/java/ru/easycode/words504/data/providers/retrofit/ProvideConverterFactory.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.data
+package ru.easycode.words504.data.providers.retrofit
 
 import retrofit2.Converter
 import retrofit2.converter.gson.GsonConverterFactory
