+++ core/src/main/java/ru/easycode/words504/data/providers/retrofit/ProvideOkHttpClientBuilder.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.data
+package ru.easycode.words504.data.providers.retrofit
 
 import okhttp3.OkHttpClient
 import java.util.concurrent.TimeUnit
