+++ core/src/main/java/ru/easycode/words504/data/providers/retrofit/ProvideRetrofitBuilder.kt
@@ -1,4 +1,4 @@
-package ru.easycode.words504.data
+package ru.easycode.words504.data.providers.retrofit
 
 import retrofit2.Retrofit
